docs: home 관련 components 분리
