+++ src/components/index.ts
@@ -1,3 +1,5 @@
 export * from "./Layout";
 export * from "./Router";
 export * from "./Menu";
+export * from "./posts/PostForm";
+export * from "./posts/PostBox";
